requirements.txt
python-telegram-bot
spamwatch
nospamplus
./modules/owlantispam
requests
pymongo
dnspython